+++ terraform/examples/ces_google_gke/ces_cluster/variables_ces.tf
@@ -71,14 +71,21 @@ variable "dogus" {
 variable "component_operator_crd_chart" {
   description = "The helm chart of the component crd. Optional with version like k8s/k8s-component-operator-crd:1.2.3"
   type        = string
-  default     = "k8s/k8s-component-operator-crd:1.10.1"
+  default     = "k8s/k8s-component-operator-crd"
 }
 
-# component operator crd
-variable "component_operator_image" {
-  description = "The Image:Version of the component operator. Optional with version like cloudogu/k8s-component-operator:1.10.0"
+# blueprint operator crd
+variable "blueprint_operator_crd_chart" {
+  description = "The helm chart of the blueprint crd. Optional with version like k8s/k8s-blueprint-operator-crd:1.2.3"
   type        = string
-  default     = "cloudogu/k8s-component-operator:1.10.1"
+  default     = "k8s/k8s-blueprint-operator-crd"
+}
+
+# ecosystem core version
+variable "ecosystem_core_chart_version" {
+  description = "The version of the ecosystem-core chart to be used. Optional as semantic version like 1.2.3"
+  type        = string
+  default     = ""
 }
 
 # List of components, backup components and monitoring components